+++ __init__.py
@@ -36,7 +36,7 @@ from scm.plams.core.settings import (
     SafeRunSettings,
     Settings,
 )
-from scm.plams.core.logging import get_logger, TextLogger
+from scm.plams.core.logging import get_logger
 from scm.plams.interfaces.adfsuite.ams import AMSJob, AMSResults
 from scm.plams.interfaces.adfsuite.amsanalysis import AMSAnalysisJob, AMSAnalysisResults, convert_to_unicode
 from scm.plams.interfaces.adfsuite.amsworker import AMSWorker, AMSWorkerError, AMSWorkerPool, AMSWorkerResults
@@ -165,7 +165,6 @@ __all__ = [
     "JobManagerSettings",
     "ConfigSettings",
     "get_logger",
-    "TextLogger",
     "SingleJob",
     "MultiJob",
     "JobStatus",